@@ -1,4 +1,4 @@
-class Constituent < ActiveRecord::Base
+class Constituent < ApplicationRecord
   belongs_to :portfolio
   belongs_to :stock
   attr_accessor :actual
@@ -10,12 +10,17 @@ class Constituent < ActiveRecord::Base
   def initialize_members
     conn = ActiveRecord::Base.connection
     @actual = self.value.to_f / get_scalar( conn, "select value from balances where type=13" ).to_f
-    if self.portfolio.id == Togabou::PORTFOLIOS_PLAY
+    if self[:portfolio] != nil && self.portfolio.id == Togabou::PORTFOLIOS_PLAY
       @actual = self.value.to_f / get_scalar( conn, "select value from balances where type=19" ).to_f
     end
 
     @off = self.actual.to_f - self.model.to_f
-    @eps_yield_wtd = self.actual.to_f * self.stock.eps_yield.to_f
-    @div_yield_wtd = self.actual.to_f * self.stock.div_yield.to_f
+    if self[:stock] != nil
+      @eps_yield_wtd = self.actual.to_f * self.stock.eps_yield.to_f
+      @div_yield_wtd = self.actual.to_f * self.stock.div_yield.to_f
+    else
+      @eps_yield_wtd = 0
+      @div_yield_wtd = 0
+    end
   end
 end
